// src/hockey/enums.ts
export enum Shoot {
  Left = 'left',
  Right = 'right',
}
export enum Backhand {
  One = 'one_hand',
  Two = 'two_hand',
}

export enum HockeyPosition {
  Defenseman = 'defenseman',
  LeftWing = 'left_wing',
  RightWing = 'right_wing',
  Center = 'center',
  Goalie = 'goalie',
}

export enum HockeyStick {
  Initial = 'initial',
  Controllers = 'controllers',
  Haptic = 'haptic',
  NonHaptic = 'non_haptic',
  None = 'none',
}